requirements.txt
# setup using:
# pip install -r requirements.txt

ruamel.yaml
munch
colorama
#termcolor